// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string | number; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: { input: any; output: any; }
  /** An object with an Global ID,for using in Noder interface. */
  GID: { input: any; output: any; }
  /** The builtin Time type */
  Time: { input: any; output: any; }
};

export type App = Node & {
  __typename?: 'App';
  actions: AppActionConnection;
  /** 应用ID */
  appKey?: Maybe<Scalars['String']['output']>;
  /** 应用密钥 */
  appSecret?: Maybe<Scalars['String']['output']>;
  /** 用于标识应用资源的唯一代码,尽量简短 */
  code: Scalars['String']['output'];
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  /** 应用类型 */
  kind: AppKind;
  /** 图标 */
  logo?: Maybe<Scalars['String']['output']>;
  menus: AppMenuConnection;
  /** 名称 */
  name: Scalars['String']['output'];
  orgs: OrgConnection;
  /** 策略 */
  policies?: Maybe<Array<AppPolicy>>;
  /** 回调地址 */
  redirectURI?: Maybe<Scalars['String']['output']>;
  /** refresh_token有效期 */
  refreshTokenValidity?: Maybe<Scalars['Int']['output']>;
  resources: AppResConnection;
  /** 角色 */
  roles?: Maybe<Array<AppRole>>;
  /** 权限范围 */
  scopes?: Maybe<Scalars['String']['output']>;
  /** 状态 */
  status?: Maybe<AppSimpleStatus>;
  /** token有效期 */
  tokenValidity?: Maybe<Scalars['Int']['output']>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AppAction = Node & {
  __typename?: 'AppAction';
  app?: Maybe<App>;
  /** 所属应用 */
  appID?: Maybe<Scalars['ID']['output']>;
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  /** restful,graphql,rpc,function */
  kind: AppActionKind;
  /** 被引用的菜单项 */
  menus?: Maybe<Array<AppMenu>>;
  /** 操作方法:读,写,列表 */
  method: AppActionMethod;
  /** 名称 */
  name: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
};

/** A connection to a list of items. */
export type AppActionConnection = {
  __typename?: 'AppActionConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<AppActionEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type AppActionEdge = {
  __typename?: 'AppActionEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<AppAction>;
};

/** AppActionKind is enum for the field kind */
export enum AppActionKind {
  Function = 'function',
  Graphql = 'graphql',
  Restful = 'restful',
  Rpc = 'rpc'
}

/** AppActionMethod is enum for the field method */
export enum AppActionMethod {
  List = 'list',
  Read = 'read',
  Write = 'write'
}
// ...
export type AppActionWhereInput = {
  and?: InputMaybe<Array<AppActionWhereInput>>;
  /** app_id field predicates */
  appID?: InputMaybe<Scalars['ID']['input']>;
  appIDIn?: InputMaybe<Array<Scalars['ID']['input']>>;
  appIDIsNil?: InputMaybe<Scalars['Boolean']['input']>;
  appIDNEQ?: InputMaybe<Scalars['ID']['input']>;
  appIDNotIn?: InputMaybe<Array<Scalars['ID']['input']>>;
  appIDNotNil?: InputMaybe<Scalars['Boolean']['input']>;
  /** created_at field predicates */
  createdAt?: InputMaybe<Scalars['Time']['input']>;
  createdAtGT?: InputMaybe<Scalars['Time']['input']>;
  createdAtGTE?: InputMaybe<Scalars['Time']['input']>;
  createdAtIn?: InputMaybe<Array<Scalars['Time']['input']>>;
  createdAtLT?: InputMaybe<Scalars['Time']['input']>;
  createdAtLTE?: InputMaybe<Scalars['Time']['input']>;
  createdAtNEQ?: InputMaybe<Scalars['Time']['input']>;
  createdAtNotIn?: InputMaybe<Array<Scalars['Time']['input']>>;
  /** created_by field predicates */
  createdBy?: InputMaybe<Scalars['Int']['input']>;
  createdByGT?: InputMaybe<Scalars['Int']['input']>;
  createdByGTE?: InputMaybe<Scalars['Int']['input']>;
  createdByIn?: InputMaybe<Array<Scalars['Int']['input']>>;
  createdByLT?: InputMaybe<Scalars['Int']['input']>;
  createdByLTE?: InputMaybe<Scalars['Int']['input']>;
  createdByNEQ?: InputMaybe<Scalars['Int']['input']>;
  createdByNotIn?: InputMaybe<Array<Scalars['Int']['input']>>;
  /** app edge predicates */
  hasApp?: InputMaybe<Scalars['Boolean']['input']>;
  hasAppWith?: InputMaybe<Array<AppWhereInput>>;
  /** menus edge predicates */
  hasMenus?: InputMaybe<Scalars['Boolean']['input']>;
  hasMenusWith?: InputMaybe<Array<AppMenuWhereInput>>;
  /** id field predicates */
  id?: InputMaybe<Scalars['ID']['input']>;
  idGT?: InputMaybe<Scalars['ID']['input']>;
  idGTE?: InputMaybe<Scalars['ID']['input']>;
  idIn?: InputMaybe<Array<Scalars['ID']['input']>>;
  idLT?: InputMaybe<Scalars['ID']['input']>;
  idLTE?: InputMaybe<Scalars['ID']['input']>;
  idNEQ?: InputMaybe<Scalars['ID']['input']>;
  idNotIn?: InputMaybe<Array<Scalars['ID']['input']>>;
  /** kind field predicates */
  kind?: InputMaybe<AppActionKind>;
  kindIn?: InputMaybe<Array<AppActionKind>>;
  kindNEQ?: InputMaybe<AppActionKind>;
  kindNotIn?: InputMaybe<Array<AppActionKind>>;
  /** method field predicates */
  method?: InputMaybe<AppActionMethod>;
  methodIn?: InputMaybe<Array<AppActionMethod>>;
  methodNEQ?: InputMaybe<AppActionMethod>;
  methodNotIn?: InputMaybe<Array<AppActionMethod>>;
  /** name field predicates */
  name?: InputMaybe<Scalars['String']['input']>;
  nameContains?: InputMaybe<Scalars['String']['input']>;
  nameContainsFold?: InputMaybe<Scalars['String']['input']>;
  nameEqualFold?: InputMaybe<Scalars['String']['input']>;
  nameGT?: InputMaybe<Scalars['String']['input']>;
  nameGTE?: InputMaybe<Scalars['String']['input']>;
  nameHasPrefix?: InputMaybe<Scalars['String']['input']>;
  nameHasSuffix?: InputMaybe<Scalars['String']['input']>;
  nameIn?: InputMaybe<Array<Scalars['String']['input']>>;
  nameLT?: InputMaybe<Scalars['String']['input']>;
  nameLTE?: InputMaybe<Scalars['String']['input']>;
  nameNEQ?: InputMaybe<Scalars['String']['input']>;
  nameNotIn?: InputMaybe<Array<Scalars['String']['input']>>;
  not?: InputMaybe<AppActionWhereInput>;
  or?: InputMaybe<Array<AppActionWhereInput>>;
  /** updated_at field predicates */
  updatedAt?: InputMaybe<Scalars['Time']['input']>;
  updatedAtGT?: InputMaybe<Scalars['Time']['input']>;
  updatedAtGTE?: InputMaybe<Scalars['Time']['input']>;
  updatedAtIn?: InputMaybe<Array<Scalars['Time']['input']>>;
  updatedAtIsNil?: InputMaybe<Scalars['Boolean']['input']>;
  updatedAtLT?: InputMaybe<Scalars['Time']['input']>;
  updatedAtLTE?: InputMaybe<Scalars['Time']['input']>;
  updatedAtNEQ?: InputMaybe<Scalars['Time']['input']>;
  updatedAtNotIn?: InputMaybe<Array<Scalars['Time']['input']>>;
  updatedAtNotNil?: InputMaybe<Scalars['Boolean']['input']>;
  /** updated_by field predicates */
  updatedBy?: InputMaybe<Scalars['Int']['input']>;
  updatedByGT?: InputMaybe<Scalars['Int']['input']>;
  updatedByGTE?: InputMaybe<Scalars['Int']['input']>;
  updatedByIn?: InputMaybe<Array<Scalars['Int']['input']>>;
  updatedByIsNil?: InputMaybe<Scalars['Boolean']['input']>;
  updatedByLT?: InputMaybe<Scalars['Int']['input']>;
  updatedByLTE?: InputMaybe<Scalars['Int']['input']>;
  updatedByNEQ?: InputMaybe<Scalars['Int']['input']>;
  updatedByNotIn?: InputMaybe<Array<Scalars['Int']['input']>>;
  updatedByNotNil?: InputMaybe<Scalars['Boolean']['input']>;
};
// ...
export enum AppKind {
  Native = 'native',
  Server = 'server',
  Web = 'web'
}

export type AppMenu = Node & {
  __typename?: 'AppMenu';
  /** 需要权限控制时对应的权限 */
  action?: Maybe<AppAction>;
  /** 操作ID */
  actionID?: Maybe<Scalars['ID']['output']>;
  app?: Maybe<App>;
  /** 所属应用 */
  appID?: Maybe<Scalars['ID']['output']>;
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  displaySort?: Maybe<Scalars['Int']['output']>;
  /** 菜单图标 */
  icon?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  /** 目录,菜单项 */
  kind: AppMenuKind;
  /** 菜单名称 */
  name: Scalars['String']['output'];
  /** 父级ID */
  parentID: Scalars['Int']['output'];
  /** 菜单路由 */
  route?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
};

/** A connection to a list of items. */
export type AppMenuConnection = {
  __typename?: 'AppMenuConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<AppMenuEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type AppMenuEdge = {
  __typename?: 'AppMenuEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<AppMenu>;
};
// ...
export type AppPolicy = Node & {
  __typename?: 'AppPolicy';
  app?: Maybe<App>;
  /** 所属应用 */
  appID?: Maybe<Scalars['ID']['output']>;
  /** 标识是否自动授予到账户 */
  autoGrant: Scalars['Boolean']['output'];
  /** 描述 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  id: Scalars['ID']['output'];
  /** 是否授权role */
  isGrantAppRole: Scalars['Boolean']['output'];
  /** 策略名称 */
  name: Scalars['String']['output'];
  roles?: Maybe<Array<AppRole>>;
  /** 策略规则 */
  rules: Array<Maybe<PolicyRule>>;
  /** 状态 */
  status?: Maybe<AppPolicySimpleStatus>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AppResConnection = {
  __typename?: 'AppResConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<AppResEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};
// ...
export type AppRole = Node & {
  __typename?: 'AppRole';
  app?: Maybe<App>;
  /** 所属应用 */
  appID?: Maybe<Scalars['ID']['output']>;
  /** 标识是否自动授予到账户 */
  autoGrant: Scalars['Boolean']['output'];
  /** 备注 */
  comments?: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Time']['output'];
  createdBy: Scalars['Int']['output'];
  /** 授权后是否可编辑 */
  editable: Scalars['Boolean']['output'];
  id: Scalars['ID']['output'];
  /** 角色名称 */
  name: Scalars['String']['output'];
  /** 权限授权策略 */
  policies?: Maybe<Array<AppPolicy>>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  updatedBy?: Maybe<Scalars['Int']['output']>;
};
// ...
export enum AppSimpleStatus {
  Active = 'active',
  Inactive = 'inactive',
  Processing = 'processing'
}
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
};
// ...
export type OrgConnection = {
  __typename?: 'OrgConnection';
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<OrgEdge>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

/** An edge in a connection. */
export type OrgEdge = {
  __typename?: 'OrgEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['Cursor']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Org>;
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['Cursor']['output']>;
  /** When paginating forwards, are there more items? */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items? */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['Cursor']['output']>;
};